Refactor MQTT (un)subscribe.
This separates the SetGatewaySubscription from the actual MQTT
(un)subscribe. When SetGatewaySubscription is called multiple times for
the same gateway while the MQTT client is disconnected, `gateways` will
always have the desired state of gateway for which a subscription must
exist. A separate handles handles the MQTT (un)subscribe and stores this
state into `gatewaysSubscribed`.

Previously when disconnected, the SetGatewaySubscription in case the
MQTT client was disconnected. When SetGatewaySubscription was called
multiple times for the same gateway to unsubscribe and subscribe, there
was a risk that it was not executed in the correct order, leaving the
gateway disconnected.

Closes #183.